refactor: batchSwap only FlowChartMixin [not working]: have rate limit issue + transfer zero amount fix: remove camelot's redeem since it's not working [fixCI]
