+++ classes/Venus/Venus.js
@@ -207,24 +207,10 @@ export class Venus extends BaseProtocol {
       assetDecimals,
     ] = this._getTheBestTokenAddressToZapOut();
 
-    const protocolContractInstance = new ethers.Contract(
-      this.protocolContract.address,
-      Vault,
-      PROVIDER(this.chain),
-    );
-
-    const balance = await protocolContractInstance.balanceOf(owner);
-    const percentageBN = ethers.utils.parseUnits(
-      percentage.toString(),
-      this.percentagePrecision,
-    );
-    const amountToWithdraw = balance
-      .mul(percentageBN)
-      .div(ethers.BigNumber.from(10).pow(this.percentagePrecision));
     const burnTxn = prepareContractCall({
       contract: this.protocolContract,
       method: "redeem",
-      params: [amountToWithdraw],
+      params: [amount],
     });
     const tradingLoss = 0;
     return [
@@ -232,7 +218,7 @@ export class Venus extends BaseProtocol {
       symbolOfBestTokenToZapInOut,
       bestTokenAddressToZapOut,
       assetDecimals,
-      amountToWithdraw,
+      amount,
       tradingLoss,
     ];
   }
